Allow group feeds to be subscribed to
`IO_MQTT#subscribe` always checks `feed_key` even though it is not required. When `group_key` is provided instead of `feed_key`, `feed_key` should not be required to be valid.

This change only validates `feed_key` when one is provided and adds validation for `group_key` when it is provided.